reservations-service/requirements.txt
Flask==2.3.3
opentelemetry-api==1.20.0
opentelemetry-distro==0.41b0
opentelemetry-exporter-otlp-proto-grpc==1.20.0
opentelemetry-sdk==1.20.0
requests==2.31.0